pagignation,limit

+++ backend/utils/removeImg.js
@@ -1,4 +1,10 @@
 import fs from "fs";
 export const removeImg = (path) => {
-  fs.unlinkSync(path);
+  try {
+    if (fs.existsSync(path)) {
+      fs.unlinkSync(path);
+    }
+  } catch (error) {
+    console.log(`file not found:${path}`);
+  }
 };
